hashing assets for cache purposes + created tabs component + organized sass
