+++ src/styles/_all.scss
@@ -1,0 +1,187 @@
+body {
+  @extend %background-color-second;
+  @extend %color-third;
+  box-sizing: border-box;
+}
+
+// Skills
+.skills .ttl {
+  text-transform: uppercase;
+  font-weight: bold;
+}
+
+section#experience,
+section#education {
+  overflow-x: hidden;
+}
+
+#only {
+  margin: 0;
+  padding: 0;
+  height: 100vh;
+}
+
+.margem-cima {
+  margin-top: 20vh;
+}
+/* ************** */
+
+/* MAIN BACKGROUND COLOR SET HERE */
+
+main {
+  @extend %background-color-main;
+  height: 100%;
+  overflow-y: scroll;
+
+  // To add strength:
+  // @TODO: find out how to override materialize grid padding settings
+  &#main {
+    padding: 0;
+  }
+}
+
+@media (max-width: 700px) {
+  #main {
+    max-width: 100vw;
+  }
+}
+
+#main-content {
+  margin: 0 10px;
+}
+
+/* ************** */
+
+/* ME SECTION */
+
+#me .row {
+  text-transform: uppercase;
+  padding: 10px;
+  padding-top: 40vh;
+  /* background-color: var(--main-fade);
+  border: 1px solid $main; */
+}
+
+.me-small {
+  text-transform: uppercase;
+  font-size: 16pt;
+}
+
+.me-big {
+  text-transform: uppercase;
+  font-size: 22pt;
+  font-weight: bold;
+  border-left: 2px #ccc solid;
+  padding-left: 5px;
+}
+
+.me-big-right {
+  text-transform: uppercase;
+  text-align: right;
+  font-size: 24pt;
+  font-weight: bold;
+  border-right: 2px #ccc solid;
+  padding-right: 5px;
+}
+
+/* ************** */
+
+/* SKILL SECTION */
+#skills {
+  padding: 5vh 0;
+}
+
+.skills-container {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: flex-start;
+}
+
+.skill {
+  display: flex;
+  justify-content: center;
+  overflow: hidden;
+  margin: 2%;
+  /* align-content: center; */
+  /* SQUARE */
+  position: relative;
+  width: 15%;
+  padding-bottom: 15%;
+  /* SQUARE PRETTY */
+  border-radius: 5px;
+  @extend %background-color-third;
+}
+
+.skill-content {
+  flex: 1;
+  position: absolute;
+  -webkit-box-sizing: border-box; /* Safari Chrome, other WebKit */
+  -moz-box-sizing: border-box; /* Firefox, other Gecko */
+  box-sizing: border-box; /* Opera/IE 8+ */
+  /* PADDING */
+  height: 90%; /* = 100% - 2*5% padding */
+  width: 90%; /* = 100% - 2*5% padding */
+  padding: 5%;
+  /* FONT */
+  font-family: "Quicksand", sans-serif;
+  font-weight: bold;
+  font-size: 50px;
+  @extend %color-main;
+  /* CENTERING */
+  display: flex;
+}
+
+@media (max-width: 700px) {
+  .skill {
+    width: 20%;
+    padding-bottom: 20%;
+  }
+  .skill-content {
+    font-size: 27px;
+  }
+}
+
+/* CENTERS IN ALL DIRECTIONS IF PARENT IS display: flex;*/
+.center-flex {
+  margin: auto;
+}
+
+.skill img {
+  width: 100%;
+  height: 100%;
+}
+
+/* ************** */
+
+/* EXPERIENCES SECTION */
+
+/* ************** */
+/* EDUCATION SECTION */
+
+/* ************** */
+/* MORE SECTION */
+#more {
+  padding-top: 5vh;
+}
+
+.spin {
+  -webkit-animation: spin 4s linear infinite;
+  -moz-animation: spin 4s linear infinite;
+  animation: spin 4s linear infinite;
+}
+@-moz-keyframes spin {
+  100% {
+    -moz-transform: rotate(360deg);
+  }
+}
+@-webkit-keyframes spin {
+  100% {
+    -webkit-transform: rotate(360deg);
+  }
+}
+@keyframes spin {
+  100% {
+    -webkit-transform: rotate(360deg);
+    transform: rotate(360deg);
+  }
+}
