--- src/styles/all.scss
@@ -1,273 +1,0 @@
-:root {
-  --main: #222430;
-  /* #282a36 */
-  --main-fade: #2a2d3c;
-  /* --second: #1c1c4a;
-  --second-light: #313181; */
-  --second: #145629;
-  --second-light: #313181;
-  --third: #eff0eb;
-  --accent-one: #339933;
-  --accent-one-dark: #1f601f;
-  --accent-two: #F7E418;
-  --accent-two-dark: #9E8C32;
-  --gray-light: #BABABA;
-}
-
-body {
-  background-color: var(--second);
-  /* FIXING SCROLL WITH NAVBAR, TEMPORARILY*/
-  color: var(--third);
-  box-sizing: border-box;
-}
-
-a {
-  color: #fff;
-}
-
-h5 {
-  text-transform: uppercase;
-  font-weight: bold;
-}
-
-section#experience, section#education {
-  overflow-x: hidden;
-}
-
-#only {
-  margin: 0;
-  padding: 0;
-}
-
-.margem-cima {
-  margin-top: 20vh;
-}
-/* ************** */
-
-/* NAVBAR PROFILE */
-
-#profile-nav {
-  background-color: var(--second);
-  position: sticky;
-  display: flex;
-  flex-direction: column;
-  padding-top: 25vh;
-  min-height: 100vh;
-  /* TEXT */
-  text-align: center;
-}
-
-@media (max-width: 700px) {
-  #profile-nav {
-    min-height: 90vh;
-    max-width: 100vw;
-  }
-}
-
-#profile-nav #info {
-  flex: 1;
-}
-
-#profile-nav footer img {
-  height: 35px;
-  width: 35px;
-  margin: 5px;
-}
-
-/* #navbar-toggle {
-  float: left;
-  position: relative;
-  left: 10px;
-  margin: auto 0;
-} */
-
-#profile-subtitle {
-  font-size: 10pt;
-  color: var(--gray-light);
-  margin: 2px;
-}
-
-#profile-name {
-  text-transform: uppercase;
-  font-weight: bold;
-  font-size: 24pt;
-}
-
-/* ************** */
-
-/* MAIN BACKGROUND COLOR SET HERE */
-
-#main {
-  background-color: var(--main);
-  /* margin-left: auto; */
-  position: static;
-  padding: 0px;
-  min-height: 100vh;
-}
-
-@media (max-width: 700px) {
-  #main {
-    max-width: 100vw;
-  }
-}
-
-#main-content {
-  margin: 0 10px;
-}
-
-/* TABS STYLING */
-.tabs-container {
-  position: static;
-  /* position: fixed;
-  max-width: 100%; */
-  background-color: var(--main-fade);
-  padding: 0 !important;
-}
-
-.tabs {
-  background-color: inherit;
-}
-
-.tabs a {
-  color: var(--third) !important;
-}
-
-.tabs a.active {
-  font-weight: bold;
-  background-color: var(--main) !important;
-  border-top: 3px solid var(--third);
-}
-
-.tabs .indicator {
-  display: none;
-  /* background-color: var(--third) !important; */
-}
-
-.tabs i.material-icons {
-  /*ADJUSTS ICONS*/
-  margin-top: 0.5em;
-}
-
-/* ************** */
-
-/* ME SECTION */
-
-#me .row {
-  text-transform: uppercase;
-  padding: 10px;
-  padding-top: 40vh;
-  /* background-color: var(--main-fade);
-  border: 1px solid var(--main); */
-}
-
-.me-small {
-  text-transform: uppercase;
-  font-size: 16pt;
-}
-
-.me-big {
-  text-transform: uppercase;
-  font-size: 22pt;
-  font-weight: bold;
-  border-left: 2px #ccc solid;
-  padding-left: 5px;
-}
-
-.me-big-right {
-  text-transform: uppercase;
-  text-align: right;
-  font-size: 24pt;
-  font-weight: bold;
-  border-right: 2px #ccc solid;
-  padding-right: 5px;
-}
-
-/* ************** */
-
-/* SKILL SECTION */
-#skills {
-  padding: 5vh 0;
-}
-
-.skills-container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-}
-
-.skill {
-  display: flex;
-  justify-content: center;
-  overflow: hidden;
-  margin: 2%;
-  /* align-content: center; */
-  /* SQUARE */
-  position: relative;
-  width: 15%;
-  padding-bottom: 15%;
-  /* SQUARE PRETTY */
-  border-radius: 5px;
-  background-color: var(--third);
-}
-
-.skill-content {
-  flex: 1;
-  position: absolute;
-  -webkit-box-sizing: border-box; /* Safari/Chrome, other WebKit */
-  -moz-box-sizing: border-box;    /* Firefox, other Gecko */
-  box-sizing: border-box;         /* Opera/IE 8+ */
-  /* PADDING */
-  height: 90%; /* = 100% - 2*5% padding */
-  width: 90%; /* = 100% - 2*5% padding */
-  padding: 5%;
-  /* FONT */
-  font-family: 'Quicksand', sans-serif;
-  font-weight: bold;
-  font-size: 50px;
-  color: var(--main);
-  /* CENTERING */
-  display: flex;
-}
-
-@media (max-width: 700px) {
-  .skill {
-    width: 20%;
-    padding-bottom: 20%;
-  }
-  .skill-content {
-    font-size: 27px;
-  }
-}
-
-/* CENTERS IN ALL DIRECTIONS IF PARENT IS display: flex;*/
-.center-flex {
-  margin: auto;
-}
-
-.skill img {
-  width: 100%;
-  height: 100%;
-}
-
-
-/* ************** */
-
-/* EXPERIENCES SECTION */
-
-/* ************** */
-/* EDUCATION SECTION */
-
-/* ************** */
-/* MORE SECTION */
-#more {
-  padding-top: 5vh;
-}
-
-.spin {
-    -webkit-animation:spin 4s linear infinite;
-    -moz-animation:spin 4s linear infinite;
-    animation:spin 4s linear infinite;
-}
-@-moz-keyframes spin { 100% { -moz-transform: rotate(360deg); } }
-@-webkit-keyframes spin { 100% { -webkit-transform: rotate(360deg); } }
-@keyframes spin { 100% { -webkit-transform: rotate(360deg); transform:rotate(360deg); } }
