+++ src/styles/components/_loader.scss
