+++ src/styles/components/_profile.scss
@@ -1,0 +1,45 @@
+/* ************** */
+
+/* NAVBAR PROFILE */
+
+#profile-nav {
+  @extend %background-color-second;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  align-items: center;
+  justify-items: center;
+}
+
+// @media (max-width: 700px) {
+//   #profile-nav {
+//     min-height: 90vh;
+//     max-width: 100vw;
+//   }
+// }
+
+#profile-nav #info {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+#profile-nav footer img {
+  height: 35px;
+  width: 35px;
+  margin: 5px;
+}
+
+#profile-subtitle {
+  font-size: 10pt;
+  color: map-get($colors, 'gray-light');
+  margin: 2px;
+}
+
+#profile-name {
+  text-transform: uppercase;
+  font-weight: bold;
+  font-size: 24pt;
+}
