+++ src/styles/components/_timeline.scss
