+++ src/styles/placeholders/_colors.scss
@@ -1,0 +1,21 @@
+$colors: (
+    'main': #222430,
+    'main-fade': #2a2d3c,
+    'second': #145629,
+    'second-light': #313181,
+    'third': #eff0eb,
+    'accent-one': #339933,
+    'accent-one-dark': #1f601f,
+    'accent-two': #F7E418,
+    'accent-two-dark': #9E8C32,
+    'gray-light': #BABABA,
+);
+
+@each $name, $color in $colors {
+    %color-#{$name} {
+        color: $color;
+    }
+    %background-color-#{$name} {
+        background: $color;
+    }
+}
