+++ tests/png/medium.png
